Analysis of data for all
