+++ Notebooks/Exploit_data.Rmd
@@ -2,7 +2,7 @@
 jupyter:
   jupytext:
     text_representation:
-      extension: .Rmd, .ipynb
+      extension: .Rmd
       format_name: rmarkdown
       format_version: '1.2'
       jupytext_version: 1.11.2
@@ -19,7 +19,7 @@ import pandas as pd
 import pickle
 import os
 
-import numpy as np 
+import numpy as np
 import matplotlib.pyplot as plt
 ```
 
@@ -29,14 +29,15 @@ import matplotlib.pyplot as plt
 ```{python}
 n_carriers_per_node = 30
 node_auction_cost = 0.
+auction_type = ['MultiLanes', 'SingleLane'][1]
 ```
 
 ```{python}
 path = os.path.join(os.getcwd(), 'Data_analysis')
 
-episodes_file_name = 'episodes_df_' + str(node_auction_cost) + '_' + str(n_carriers_per_node) + '.bin'
-loads_file_name = 'loads_df_' + str(node_auction_cost) + '_' + str(n_carriers_per_node) + '.bin'
-movements_file_name = 'movements_df_' + str(node_auction_cost) + '_' + str(n_carriers_per_node) + '.bin'
+episodes_file_name = 'episodes_df_' + auction_type + '_' + str(node_auction_cost) + '_' + str(n_carriers_per_node) + '.bin'
+loads_file_name = 'loads_df_' + auction_type + '_' + str(node_auction_cost) + '_' + str(n_carriers_per_node) + '.bin'
+movements_file_name = 'movements_df_' + auction_type + '_' + str(node_auction_cost) + '_' + str(n_carriers_per_node) + '.bin'
 
 with open(os.path.join(path, episodes_file_name), 'rb') as f:
     episodes_df = pickle.load(f)
@@ -222,4 +223,8 @@ hops_per_departure_destination = loads_df_filtered.groupby(['departure', 'arriva
 
 print(price_per_departure_destination)
 print(hops_per_departure_destination)
+```
+
+```{python}
+
 ```
