+++ Notebooks/Get_exploitable_data.Rmd
@@ -24,7 +24,7 @@ import os
 
 # Objects
 from PI_RPS.Mechanics.Environment.environment import Environment
-from PI_RPS.Mechanics.Actors.Carriers.cost_bidding_carrier import CostBiddingCarrier
+from PI_RPS.Mechanics.Actors.Carriers.cost_bidding_carrier import MultiLanesCostBiddingCarrier, SingleLaneCostBiddingCarrier
 
 # Tools
 from PI_RPS.Games.init_tools import load_realistic_nodes_and_shippers_to_env
@@ -121,8 +121,7 @@ def new_environment():
                                          transit_cost=road_costs,
                                          far_from_home_cost=drivers_costs,
                                          time_not_at_home=0,
-                                         max_time_not_at_home=max_time_not_at_home,
-                                         too_high_bid=shipper_default_reserve_price)
+                                         max_time_not_at_home=max_time_not_at_home)
         else:
             raise NotImplementedError
     return e
